requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
bcrypt==3.1.7
better-profanity
Cerberus==1.3.2
colorama==0.4.3
Django==3.0.7
django-cors-headers==3.2.1
django-easy-validator==1.4.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.20
mysqlclient==1.4.6
protobuf==3.12.2
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
PyYAML==5.4
pylint==2.5.3
PyMySQL==0.9.3
python-http-client==3.2.7
pytz==2020.1
sendgrid==6.4.1
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
Werkzeug
cffi
redis
pyfcm==1.4.7
boto3
apscheduler
cryptography==3.3.1
deepdiff==5.2.2
